backend/requirements.txt
fastapi
uvicorn
pystac-client
requests
rasterio
numpy
matplotlib
Pillow
torch
segmentation-models-pytorch
pydantic
